requirements.txt
asgiref==3.8.1
black==24.4.2
click==8.1.7
colorama==0.4.6
Django>=4.0.4
djangorestframework==3.13.1
flake8==5.0.4
flake8-quotes==3.3.1
flake8-variables-names==0.0.5
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.0
pathspec==0.12.1
pep8-naming==0.13.2
platformdirs==4.2.2
pycodestyle==2.9.1
pyflakes==2.5.0
pytz==2024.1
sqlparse==0.5.0
tzdata==2024.1
